requirements.txt
Pillow>=9.0.0
flask>=2.0.0
requests>=2.25.0
pyyaml>=6.0.0
tabulate>=0.9.0
flask-cors>=4.0.0
